move typeck constraints conversion to its own module
